ログの初期化をエントリーポイントに移動
