+++ src/Beutl/App.axaml.cs
@@ -19,11 +19,7 @@ using Beutl.Views;
 using FluentAvalonia.Core;
 using FluentAvalonia.Styling;
 
-using Microsoft.Extensions.Logging;
-
 using Reactive.Bindings;
-
-using Serilog;
 
 namespace Beutl;
 
@@ -34,8 +30,6 @@ public sealed class App : Application
 
     public override void Initialize()
     {
-        SetupLogger();
-
         FAUISettings.SetAnimationsEnabledAtAppLevel(true);
 
         //PaletteColors
@@ -89,24 +83,6 @@ public sealed class App : Application
         LibraryRegistrar.RegisterAll();
         NodesRegistrar.RegisterAll();
         ReactivePropertyScheduler.SetDefault(AvaloniaScheduler.Instance);
-    }
-
-    private static void SetupLogger()
-    {
-        string logFile = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".beutl", "log", "log.txt");
-        const string OutputTemplate = "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] [{SourceContext:l}] {Message:lj}{NewLine}{Exception}";
-        Log.Logger = new LoggerConfiguration()
-            .Enrich.FromLogContext()
-#if DEBUG
-            .MinimumLevel.Verbose()
-            .WriteTo.Debug(outputTemplate: OutputTemplate)
-#else
-            .MinimumLevel.Debug()
-#endif
-            .WriteTo.Async(b => b.File(logFile, outputTemplate: OutputTemplate, shared: true, rollingInterval: RollingInterval.Day))
-            .CreateLogger();
-
-        BeutlApplication.Current.LoggerFactory = LoggerFactory.Create(builder => builder.AddSerilog(Log.Logger, true));
     }
 
     public override void OnFrameworkInitializationCompleted()
